Issue 4657 fix: Changed to Double from Integer

+++ usermanagement-services-lib/src/main/java/org/ekstep/genieservices/profile/SummarizerServiceImpl.java
@@ -170,7 +170,7 @@ public class SummarizerServiceImpl extends BaseService implements ISummarizerSer
             if (learnerAssessmentDetailsModel1 == null) {
                 response.setResult(questionReportMapList);
             } else {
-                Map<Integer, Integer> accuracyReport = learnerAssessmentDetailsModel1.getAccuracyReportMap();
+                Map<Double, Integer> accuracyReport = learnerAssessmentDetailsModel1.getAccuracyReportMap();
 
                 if (accuracyReport == null) {
                     response.setResult(questionReportMapList);
